refactor search input
